Move callable mobpack execution into mob runtime
