+++ meerkat-mob/Cargo.toml
@@ -49,6 +49,7 @@ jsonschema = { workspace = true }
 futures = { workspace = true }
 tokio-util = { workspace = true }
 sha2 = { workspace = true }
+schemars = { workspace = true, optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 tokio = { workspace = true }
@@ -94,6 +95,7 @@ default = ["runtime-adapter"]
 # is empty because the dep is always present.
 runtime-adapter = []
 integration-real-tests = []
+schema = ["dep:schemars"]
 test-support = []
 
 [package.metadata.cargo-machete]
